Move MintRow location
