+++ components/Layout/MintRow.tsx
