+++ components/Layout/index.ts
@@ -1,1 +1,2 @@
 export { default as Wallet } from "./Wallet";
+export { default as MintRow } from "./MintRow";
